[STEP 07, 08] 서버 구축-데이터베이스 기본 (#24)
* fix(ControllerAdvise): GlobalExceptionHandler를 ResponseEntityExceptionHandler 기반으로 개선

* fix(config): common 패키지에서 분리

> 스프링 프레임워크에 의존적인 설정/config 등의

* fix(ControllerAdvice): common 패키지에서 분리

* doc(Controller): URI 설계 원칙 및 예시 보완

* fix(exception): service.common 패키지에서 분리

* fix(transaction): service.common 패키지에서 분리

* chore(service): service.common 패키지 삭제됨에 따른 패키지 임포트 정리

* chore(test): common 패키지 삭제됨에 따른 패키지 임포트 정리

* fix(Transaction): CompensationScope.executeWithCompensation

> 사용처 코드가 너무 장황해져서 try/catch 로직을 CompensationScope.kt 내부로 캡술화

* fix(User): UserService.kt 인터페이스 추출 및 도메인 구현 수정

> User.requiresId() 캡슐화
> Port, Usecase 메서드명을 `find~` 패턴으로 변경

* fix(Product): ProductService.kt 인터페이스 추출 및 도메인 구현 수정

> Product.requiresId() 캡슐화
> Product.id를 nullable로 변경
> 명명된 매개변수 적극 도입

* fix(Point): PointService.kt 인터페이스 추출 및 도메인 구현 수정

> PointChange.id를 nullable로 변경
> 명명된 매개변수 적극 도입

* fix(Coupon): CouponService.kt 인터페이스 추출 및 도메인 구현 수정

> Coupon.id, UserCoupon.id를 nullable로 변경
> UserCoupon.isUsable() 캡슐화
> 조회 메서드 명을 `find~` 패턴으로 변경
> 명명된 파라미터 적극 도입

> 명명된 매개변수 적극 도입

* fix(Order): OrderService.kt 인터페이스 추출 및 도메인 구현 수정

> Order.id, OrderItem.id를 nullable로 변경
> 주문 프로세스를 단계별로 메서드 분리
> 1. 주문 생성 맥락에 필요한 객체 로딩
> 2. 주문 생성 및 보상 트랜잭션 처리
> 명명된 매개변수 적극 도입

* fix(User): 임시 Adapter 구현체 상속 메서드명 수정

* fix(Controller): 컨트롤러 레이어 전체적 개선

> URI 일관성있게 수정 (/coupons)
> 서비스 대신 유스케이스 인터페이스를 주입받도록 수정
> 명명된 파라미터를 최대한 도입

* test(Service): 서비스 테스트 코드 수정

> Usecase 기반 코드베이스 수정 반영
> 명명된 파라미터를 최대한 도입

* fix(Controller): 페이지네이션 대응, 리소스 생성 반환을 201로 변경, 코루틴 의존성 제거

> 코루틴 자꾸 처리 중에 error 뜨고 모델 바인딩이 안되서 runBlocking 처리함..
> 페이지네이션 진작할껄 ㅠㅠ

* feat(Pagination): 페이지네이션 관련 공통 기능 추가

> PagingOptions.kt
> PagedList.kt

* fix(Product): 페이지네이션 대응

* fix(Coupon): 페이지네이션 대응

* doc(service): common 분류 삭제 문서 최신화

* fix(exception): exception throw 패턴 일관성 작업

* fix(Coupon): 페이지네이션 대응

* fix(Coupon): 쿠폰 회수 시 시간 입력 받도록 수정

* fix(Coupon): 쿠폰 발급 시 시간 입력 받도록 수정

* feat(Coupon): 사용 불가능한 쿠폰 예외

* fix(service): 일관적으로 입,출력에서 Long 사용하도록 변경

* feat(TimeProvider): Instant 대응 추가

* feat(Unwrapper): Optional<T> 언래핑 헬퍼

* chore: springdoc 버전 최신화

* [STEP 07, 08] 서버 구축-데이터베이스 기본 (#19)

* chore: JPA 자동 Repository 빈 주입 설정

* feat(JPA Entity): JPA DB Entity 작업

* feat(JPA Repository): JPA Repository 작업

* chore(QueryDSL): QueryDSL 의존성 설정

* feat(PopularProduct): 인기 상품 조회 Repository 작업

* feat(Coupon): 쿠폰 Adapter

* feat(Order): 주문 Adapter, 데이터 수집 더미 Adapter

* feat(Point): 포인트 Adapter

* feat(Product): 상품 Adapter

* feat(User): 회원 Adapter

* doc(Repository): repository 레이어 개발 지침 문서 최신화

* chore(QueryDSL): QueryDSL 빌드 설정

* chore(spy): SQL 로그 분석을 위한 밑작업

> 로그 세팅
> p6spy 세팅

* doc(dummy data): 회원 더미 데이터 SQL

* doc(dummy data): 상품 더미 데이터 SQL

* doc(dummy data): 쿠폰 더미 데이터 SQL

* doc(dummy data): 주문 및 주문 상품 더미 데이터 SQL

* WIP: 쿼리 병목 지점 분석 작업 진행 중

* [STEP 07, 08] 피드백 반영 (#22)

* fix(Order): 외부 데이터 플랫폼 전송 로직을 보상 트랜잭션에 포함

* refactor(Order): 외부 도메인 의존성을 facade로 분리

* test(Service): PagedList 기반 리팩터를 테스트에 적용

* test(Controller): WebMvcTest & RestClient 기반 컨트롤러 테스트로 리팩터

* fix(Controller): Valid 어노테이션 적용

* chore(Validation): Valid 어노테이션 적용 의존성 세팅

* doc(Test): 테스트 패키지 README.md 및 instruction.md 적용

* test(Integration): 테스트컨테이너 테스트 예시 추가

* test(Transaction): CompensationScope.kt 동작 테스트케이스

* test(Rollback): 쿠폰, 주문 롤백 로직 테스트 케이스 추가

+++ src/main/kotlin/kr/hhplus/be/server/service/product/entity/ProductSaleSummary.kt
@@ -5,7 +5,7 @@ import java.time.ZonedDateTime
 data class ProductSaleSummary(
     val product: Product,
     val rank: Int,
-    val soldCount: Int,
+    val soldCount: Long,
     val from: ZonedDateTime,
     val until: ZonedDateTime
 )
